requirements.txt
amqp==2.6.0
asgiref==3.2.10
asyncio==3.4.3
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.8
django-jsonfield==1.4.0
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
psycopg2-binary==2.8.5
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
zipp==3.1.0
